Fix clang [-Wpointer-sign] warnings
libnet_build_icmpv4_echo() expects the padding parameter to be a pointer
of type const uint8_t*, but in pingmac_send() a const char* is passed,
causing clang to emit a [-Wpointer-sign] warning.

The same is the case for the padding parameter to libnet_build_arp used
in pingip_send().

Fix these two warnings by defining the padding as const uint8_t[16].

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>