Fix false duplicates in edge condition
We still have duplicates when pinging in promisc mode, if someone else
is arping at the same time. It's assumed that arping in promisc mode
is rare, and that in either case what we're interested in then is the
reply, no matter what the reason was for it being sent.

This is to solve a problem where even if not using promisc mode we'll
see replies going *out* of localhost. This fix assumes that default
desired behaviour when arpinging local host is to not get replies from
local host, only from remote hosts.

Default behavour, after this commit, uses the model of request and
reply as best it can (localhost will never request from itself, and
will actually not even reply to it). It therefore doesn't work well if
the *actual* question is "is this IP address taken", because if the
answer is "yes, by localhost" the answer is now "no", whereas before
this commit it was *unreliably* sometimes "yes".

Fixes #32