Also try lo0 as a fallback interface. Fixes default use on OpenBSD
Background:

Libnet usually needs init before we have searched for the real
interface. In that case, first we just give a NULL pointer as the
interface. But libnet sometimes fails to find an interface (no idea
why), so we try to put in "lo".

This commit adds "lo0" as an interface checked, as that is the name on
OpenBSD.