Don't attempt to use caps if header doesn't exist
Termux with android seems to not provide capabilities.h, even though
the libc provides it.

Without the header it's not actually possible to use it though, so
don't even try.

There's a similar situation with getrandom(), but it's still possible
to call it without a prototype, so I'm opting to just let that spit
out a warning at build time on those systems.

Bug #49