Add seccomp to drop syscall access before receiving any packets
seccomp is not as good as pledge(), in that different systems, and
even different versions of any of the transitively dependent
libraries, will need different syscalls.

This feature is therefore off by default, and should be considered
experimental.

Maybe the right long term solution is to blacklist, not
whitelist. Which is also not great, as new syscalls get created.