configure.ac: make --disable-seccomp work
The first branch ("[action-if-given]") is taken even if --disable-seccomp
is passed. So, in that branch, check whether the user disabled it or not.

Without this, we'd get a failure when seccomp is detected but we want
to disable it:
```
./configure --disable-seccomp
[...]
checking for library containing clock_gettime... none required
configure: error: --enable-seccomp given but seccomp libraries not present
```

Signed-off-by: Sam James <sam@gentoo.org>